imp: Crosspost and save posts to database at same time
Posts are now saved to the database as they are crossposted (previously
it would crosspost every post before inserting all posts into the
database as a single transaction). This approach means that the user can
safely restart the program and it will resume from the last post the was
crossposted.

Each post is saved to the database as a separate transation transaction
meaning there is a reduced throughput when compared to a bulk insert.
However, the time it takes to create a post on Reddit is still much
higher than saving a post to the DB. And as posts are saved and
x-posted using non-blocking async calls, the post can be saved to
the database while waiting for the Reddit API to respond.

If the "EnableCrossposting" setting is false, the program will save the
posts to the database as a single bulk insert to improve performance.